refactor(aggregator) move HumanReadableProtocolConfiguration into tools
